add package libs/http-utils
