+++ gateway/Cargo.toml
@@ -4,22 +4,23 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-tokio = { version = "1.36", features = ["full"] }
-axum = { version = "0.7.4", features = ["multipart", "tracing"] }
-tower-http = { version = "0.5.2", features = ["fs", "auth", "trace", "cors"] }
-anyhow = { version = "1", features = ["backtrace"] }
+http-utils = { path = "../libs/http-utils" }
 signal = { path = "../libs/signal" }
 live777-http = { path = "../libs/live777-http" }
 utils = { path = "../libs/utils" }
+
+axum = { version = "0.7.4", features = ["multipart", "tracing"] }
+tower-http = { version = "0.5.2", features = ["fs", "auth", "trace", "cors"] }
+anyhow = { version = "1", features = ["backtrace"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 clap = { version = "4.5.1", features = ["derive"] }
 serde = { workspace = true, features = ["serde_derive"] }
 base64 = "0.22.1"
+tokio = { version = "1.36", features = ["full"] }
 toml = "0.8.10"
 http = "1.0.0"
 http-body = "1.0.0"
-http-body-util = "0.1.0"
 hyper = { version = "1.0.0", features = ["full"] }
 hyper-util = { version = "0.1.1", features = ["client-legacy"] }
 url = "2.5"
