+++ gateway/src/main.rs
@@ -1,16 +1,9 @@
 use crate::route::r#static::static_server;
-use axum::body::{Body, Bytes};
+use axum::body::Body;
 use axum::extract::Request;
-use axum::http::HeaderMap;
-use axum::middleware::Next;
-use axum::{
-    http::StatusCode,
-    response::{IntoResponse, Response},
-    Router,
-};
+use axum::Router;
 use clap::Parser;
 
-use http_body_util::BodyExt;
 use hyper_util::client::legacy::connect::HttpConnector;
 use hyper_util::rt::TokioExecutor;
 use sqlx::mysql::MySqlConnectOptions;
@@ -83,7 +76,7 @@ async fn main() {
         } else {
             CorsLayer::new()
         })
-        .layer(axum::middleware::from_fn(print_request_response))
+        .layer(axum::middleware::from_fn(http_utils::print_request_response))
         .layer(
             TraceLayer::new_for_http().make_span_with(|request: &Request<_>| {
                 let span = info_span!(
@@ -112,48 +105,4 @@ struct AppState {
     config: Config,
     pool: MySqlPool,
     client: Client,
-}
-
-async fn print_request_response(
-    req: Request,
-    next: Next,
-) -> std::result::Result<impl IntoResponse, (StatusCode, String)> {
-    let req_headers = req.headers().clone();
-    let (parts, body) = req.into_parts();
-    let bytes = buffer_and_print("request", req_headers, body).await?;
-    let req = Request::from_parts(parts, Body::from(bytes));
-
-    let res = next.run(req).await;
-    let res_headers = res.headers().clone();
-    let (parts, body) = res.into_parts();
-    let bytes = buffer_and_print("response", res_headers, body).await?;
-    let res = Response::from_parts(parts, Body::from(bytes));
-
-    Ok(res)
-}
-
-async fn buffer_and_print<B>(
-    direction: &str,
-    headers: HeaderMap,
-    body: B,
-) -> std::result::Result<Bytes, (StatusCode, String)>
-where
-    B: axum::body::HttpBody<Data = Bytes>,
-    B::Error: std::fmt::Display,
-{
-    let bytes = match body.collect().await {
-        Ok(collected) => collected.to_bytes(),
-        Err(err) => {
-            return Err((
-                StatusCode::BAD_REQUEST,
-                format!("failed to read {direction} body: {err}"),
-            ));
-        }
-    };
-
-    if let Ok(body) = std::str::from_utf8(&bytes) {
-        debug!("{direction} headers = {headers:?} body = {body:?}");
-    }
-
-    Ok(bytes)
 }
