+++ libs/http-utils/src/lib.rs
@@ -1,0 +1,54 @@
+use axum::body::{Body, Bytes};
+use axum::extract::Request;
+use axum::http::HeaderMap;
+use axum::middleware::Next;
+use axum::{
+    http::StatusCode,
+    response::{IntoResponse, Response},
+};
+use http_body_util::BodyExt;
+use tracing::debug;
+
+pub async fn print_request_response(
+    req: Request,
+    next: Next,
+) -> std::result::Result<impl IntoResponse, (StatusCode, String)> {
+    let req_headers = req.headers().clone();
+    let (parts, body) = req.into_parts();
+    let bytes = buffer_and_print("request", req_headers, body).await?;
+    let req = Request::from_parts(parts, Body::from(bytes));
+
+    let res = next.run(req).await;
+    let res_headers = res.headers().clone();
+    let (parts, body) = res.into_parts();
+    let bytes = buffer_and_print("response", res_headers, body).await?;
+    let res = Response::from_parts(parts, Body::from(bytes));
+
+    Ok(res)
+}
+
+pub async fn buffer_and_print<B>(
+    direction: &str,
+    headers: HeaderMap,
+    body: B,
+) -> std::result::Result<Bytes, (StatusCode, String)>
+where
+    B: axum::body::HttpBody<Data = Bytes>,
+    B::Error: std::fmt::Display,
+{
+    let bytes = match body.collect().await {
+        Ok(collected) => collected.to_bytes(),
+        Err(err) => {
+            return Err((
+                StatusCode::BAD_REQUEST,
+                format!("failed to read {direction} body: {err}"),
+            ));
+        }
+    };
+
+    if let Ok(body) = std::str::from_utf8(&bytes) {
+        debug!("{direction} headers = {headers:?} body = {body:?}");
+    }
+
+    Ok(bytes)
+}
