+++ liveion/src/lib.rs
@@ -1,23 +1,15 @@
-use axum::body::{Body, Bytes};
 use axum::extract::Request;
-use axum::http::HeaderMap;
-use axum::middleware::Next;
 use axum::routing::get;
-use axum::{
-    http::StatusCode,
-    response::{IntoResponse, Response},
-    Router,
-};
+use axum::Router;
 
 use error::AppError;
-use http_body_util::BodyExt;
 use std::future::Future;
 use std::sync::Arc;
 use tokio::net::TcpListener;
 use tower_http::cors::CorsLayer;
 use tower_http::trace::TraceLayer;
 use tower_http::validate_request::ValidateRequestHeaderLayer;
-use tracing::{debug, error, info_span};
+use tracing::{error, info_span};
 
 use crate::auth::ManyValidate;
 use crate::config::Config;
@@ -66,7 +58,7 @@ where
         } else {
             CorsLayer::new()
         })
-        .layer(axum::middleware::from_fn(print_request_response))
+        .layer(axum::middleware::from_fn(http_utils::print_request_response))
         .layer(
             TraceLayer::new_for_http().make_span_with(|request: &Request<_>| {
                 let span = info_span!(
@@ -109,48 +101,4 @@ async fn metrics() -> String {
     metrics::ENCODER
         .encode_to_string(&metrics::REGISTRY.gather())
         .unwrap()
-}
-
-async fn print_request_response(
-    req: Request,
-    next: Next,
-) -> std::result::Result<impl IntoResponse, (StatusCode, String)> {
-    let req_headers = req.headers().clone();
-    let (parts, body) = req.into_parts();
-    let bytes = buffer_and_print("request", req_headers, body).await?;
-    let req = Request::from_parts(parts, Body::from(bytes));
-
-    let res = next.run(req).await;
-    let res_headers = res.headers().clone();
-    let (parts, body) = res.into_parts();
-    let bytes = buffer_and_print("response", res_headers, body).await?;
-    let res = Response::from_parts(parts, Body::from(bytes));
-
-    Ok(res)
-}
-
-async fn buffer_and_print<B>(
-    direction: &str,
-    headers: HeaderMap,
-    body: B,
-) -> std::result::Result<Bytes, (StatusCode, String)>
-where
-    B: axum::body::HttpBody<Data = Bytes>,
-    B::Error: std::fmt::Display,
-{
-    let bytes = match body.collect().await {
-        Ok(collected) => collected.to_bytes(),
-        Err(err) => {
-            return Err((
-                StatusCode::BAD_REQUEST,
-                format!("failed to read {direction} body: {err}"),
-            ));
-        }
-    };
-
-    if let Ok(body) = std::str::from_utf8(&bytes) {
-        debug!("{direction} headers = {headers:?} body = {body:?}");
-    }
-
-    Ok(bytes)
 }
